requirements.txt
addict==2.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.3.0
audioread==3.0.1
av==15.1.0
babel==2.17.0
beautifulsoup4==4.13.5
billiard==4.2.3
blis==1.3.0
boto3==1.28.60
botocore==1.31.85
Brotli==1.1.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.22.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
csvw==3.6.0
ctranslate2==4.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
decorator==5.2.1
Django==4.2.13
django-environ==0.12.0
django-storages==1.14.6
django_celery_results==2.6.0
dlinfo==2.0.0
docopt==0.6.2
EbookLib==0.19
espeakng-loader==0.2.4
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
gunicorn==23.0.0
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kokoro==0.9.4
kokoro-onnx==0.3.9
kokoro-tts==2.3.0
kombu==5.5.4
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
loguru==0.7.3
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
mutagen==1.47.0
networkx==3.5
num2words==0.5.14
numba==0.62.0
numpy==2.3.3
onnxruntime==1.23.0
packaging==25.0
phonemizer-fork==3.3.1
Pillow==10.0.0
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.10
prompt_toolkit==3.0.52
protobuf==6.32.1
psycopg2-binary==2.9.10
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis>=5.0.0
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.31.0
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.1
s3transfer==0.7.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
segments==2.3.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stripe==12.4.0
sympy==1.14.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
websockets==15.0.1
wheel==0.45.1
whitenoise==6.11.0
win32_setctime==1.2.0
wrapt==1.17.3
yt-dlp==2025.11.12
dj-database-url>=2.1.0